Frontend con Vue.js

+++ Backend/Cartera-Cripto-Api/Models/Transaccion.cs
@@ -16,9 +16,10 @@ namespace Cartera_Cripto.Models
         public string crypto_code { get; set; }
 
         [Required]
-        public float crypto_amount { get; set; }
+        public decimal crypto_amount { get; set; }
 
-        public float money { get; set; }
+        public decimal money { get; set; }
+
         public DateTime datetime { get; set; }
 
         [Required]
